Remove old swc-plugin, rename swc-plugin2 to swc-plugin, update references and metadata; fix tests to ES imports
