+++ packages/parcel-transformer-swc/README.md
@@ -1,7 +1,5 @@
 @compiled/parcel-transformer-swc
 
-Parcel transformer that applies @compiled/swc-plugin2 and feeds extracted CSS to @compiled/parcel-optimizer via asset.meta.styleRules.
+Parcel transformer that applies @compiled/swc-plugin and feeds extracted CSS to @compiled/parcel-optimizer via asset.meta.styleRules.
 
 Usage: extend .parcelrc and add "@compiled/parcel-transformer-swc" before default JS transformers.
-
-
