+++ packages/swc-plugin/jest.config.js
@@ -5,7 +5,7 @@ module.exports = {
   rootDir: __dirname,
   testEnvironment: 'node',
   setupFilesAfterEnv: ['jest-extended'],
-  testMatch: ['**/tests/**/*.test.{js,ts,tsx}', '**/tests/**/*.js'],
+  testMatch: ['**/tests/**/*.test.{js,ts,tsx}'],
   moduleNameMapper: {
     '^@compiled/react/runtime$': '<rootDir>/../react/src/runtime.ts',
     '^@compiled/react/(.*)$': '<rootDir>/../react/src/jsx/$1.ts',
