+++ packages/swc-plugin/src/utils/mod.rs
@@ -1,2 +1,4 @@
 pub mod ast;
 pub mod css_builder;
+pub mod xcss;
+
