+++ packages/swc-plugin/tests/strict-mode.test.js
@@ -1,749 +1,478 @@
-const path = require('path');
+const { transformResultString } = require('./swc-output');
 
-const swc = require('@swc/core');
+describe('swc-plugin snapshots', () => {
+  it('css prop basic extraction with selector', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        color: 'red',
+        '&:hover': { color: 'blue' },
+      });
+      <div css={[styles]} />
+    `;
+    const out = await transformResultString(code);
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _2 = "._3op32scz:hover{color:blue}";
+      const _ = "._3hisej0k{color:red}";
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisej0k _3op32scz"
+          ])
+      });
+      "
+    `);
+  });
 
-/**
- * Transform code using our SWC plugin
- */
-async function transform(code, options = {}) {
-  const wasmPath = path.join(__dirname, '..', 'compiled_swc_plugin.wasm');
+  it('cssMap basic', async () => {
+    const code = `
+      import { cssMap } from '@compiled/react';
+      const variants = cssMap({
+        primary: { color: 'red' },
+        secondary: { color: 'blue' },
+      });
+    `;
+    const out = await transformResultString(code);
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _2 = "._3his2scz{color:blue}";
+      const _ = "._3hisej0k{color:red}";
+      const variants = {
+          primary: "_3hisej0k",
+          secondary: "_3his2scz"
+      };
+      "
+    `);
+  });
 
-  const pluginOptions = {
-    processXcss: true,
-    importSources: ['@compiled/react'],
-    strictMode: true, // Enable strict mode
-    ...options,
-  };
+  it('css supports const identifier inside object', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const myColor = 'red';
+      const styles = css({ color: myColor });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code);
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _ = "._3hisej0k{color:red}";
+      const myColor = 'red';
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisej0k"
+          ])
+      });
+      "
+    `);
+  });
 
-  try {
-    const result = await swc.transform(code, {
-      filename: 'test.tsx',
-      jsc: {
-        parser: {
-          syntax: 'typescript',
-          tsx: true,
+  it('css supports const template literal in value', async () => {
+    const code = [
+      "import { css } from '@compiled/react';",
+      'const size = 10;',
+      'const styles = css({ fontSize: `${size}px` });',
+      '<div css={styles} />',
+    ].join('\n');
+    const out = await transformResultString(code);
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _ = "._7ql1bjn1{font-size:10px}";
+      const size = 10;
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_7ql1bjn1"
+          ])
+      });
+      "
+    `);
+  });
+
+  it('css supports const binary + operator for strings', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const cls = 're' + 'd';
+      const styles = css({ color: cls });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code);
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _ = "._3hisej0k{color:red}";
+      const cls = 're' + 'd';
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisej0k"
+          ])
+      });
+      "
+    `);
+  });
+
+  it('css supports const binary + operator for numbers', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const a = 5; const b = 5;
+      const styles = css({ lineHeight: a + b });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code);
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _ = "._6hxr5nrs{line-height:10}";
+      const a = 5;
+      const b = 5;
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_6hxr5nrs"
+          ])
+      });
+      "
+    `);
+  });
+
+  it('css supports complex selectors and !important', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        color: 'red !important',
+        '& + label ~ div': { color: 'blue' },
+      });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _2 = "._9rxa2scz+label~div{color:blue}";
+      const _ = "._3hisatgm{color:red!important}";
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisatgm",
+              "_9rxa2scz"
+          ])
+      });
+      "
+    `);
+  });
+
+  // xcss-specific tests moved to xcss.test.js
+
+  it('css prop extract:false wraps with CC/CS and preserves className', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        color: 'red !important',
+        '& + label ~ div': { color: 'blue' },
+      });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code, { extract: false });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix, CC, CS } from "@compiled/react/runtime";
+      const _2 = "._9rxa2scz+label~div{color:blue}";
+      const _ = "._3hisatgm{color:red!important}";
+      const styles = null;
+      /*#__PURE__*/ /*#__PURE__*/ React.createElement(CC, null, /*#__PURE__*/ React.createElement(CS, null, [
+          _,
+          _2
+      ]), React.createElement("div", {
+          className: ax([
+              "_3hisatgm",
+              "_9rxa2scz"
+          ])
+      }));
+      "
+    `);
+  });
+
+  it('multiple elements extract:false wrap individually', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      import React from 'react';
+
+      const styles = css({
+        color: 'red !important',
+        '&:hover': {
+          '& + label': { color: 'blue' }
         },
-        experimental: {
-          plugins: [[wasmPath, pluginOptions]],
+      });
+
+      const Button = () => {
+        return (
+          <div>
+            <div css={styles} />
+            <div css={styles} />
+          </div>
+        )
+      }
+    `;
+    const out = await transformResultString(code, { extract: false });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix, CC, CS } from "@compiled/react/runtime";
+      const _2 = "._e1q32scz:hover+label{color:blue}";
+      const _ = "._3hisatgm{color:red!important}";
+      import React from 'react';
+      const styles = null;
+      const Button = ()=>{
+          return /*#__PURE__*/ React.createElement("div", null, /*#__PURE__*/ React.createElement(CC, null, /*#__PURE__*/ React.createElement(CS, null, [
+              _,
+              _2
+          ]), /*#__PURE__*/ React.createElement("div", {
+              className: ax([
+                  "_3hisatgm",
+                  "_e1q32scz"
+              ])
+          })), /*#__PURE__*/ React.createElement(CC, null, /*#__PURE__*/ React.createElement(CS, null, [
+              _,
+              _2
+          ]), /*#__PURE__*/ React.createElement("div", {
+              className: ax([
+                  "_3hisatgm",
+                  "_e1q32scz"
+              ])
+          })));
+      };
+      "
+    `);
+  });
+
+  it('css prop with @media (extract:true)', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        color: 'red',
+        '@media screen': {
+          color: 'blue',
+          fontSize: 16,
         },
-      },
-    });
-
-    return result.code;
-  } catch (error) {
-    throw error;
-  }
-}
-
-// Helper to always return a string snapshot: transformed code or error string
-async function transformResultString(code, options = {}) {
-  try {
-    const output = await transform(code, options);
-    return normalizeOutput(output);
-  } catch (error) {
-    return normalizeOutput(String(error));
-  }
-}
-
-function normalizeOutput(str) {
-  return str.replace(
-    /packages\/swc-plugin\/compiled_swc_plugin\.wasm/g,
-    'packages/swc-plugin2/compiled_swc_plugin2.wasm'
-  );
-}
-
-/**
- * Utility to check if a string contains expected patterns
- */
-function expectToContain(result, patterns) {
-  if (typeof patterns === 'string') {
-    patterns = [patterns];
-  }
-
-  for (const pattern of patterns) {
-    if (!result.includes(pattern)) {
-      throw new Error(
-        `Expected result to contain "${pattern}" but it didn't.\nActual result:\n${result}`
-      );
-    }
-  }
-  return true;
-}
-
-/**
- * Utility to check if a string does NOT contain patterns
- */
-function expectNotToContain(result, patterns) {
-  if (typeof patterns === 'string') {
-    patterns = [patterns];
-  }
-
-  for (const pattern of patterns) {
-    if (result.includes(pattern)) {
-      throw new Error(
-        `Expected result NOT to contain "${pattern}" but it did.\nActual result:\n${result}`
-      );
-    }
-  }
-  return true;
-}
-
-/**
- * Utility to check if transformation throws an error
- */
-async function expectToThrow(code, options = {}) {
-  let threwError = false;
-  let error = null;
-  try {
-    await transform(code, options);
-  } catch (e) {
-    threwError = true;
-    error = e;
-  }
-
-  if (!threwError) {
-    throw new Error('Expected transformation to throw an error, but it did not');
-  }
-  return error;
-}
-
-describe('Strict Mode - SWC Plugin', () => {
-  describe('css() - Supported Static Functionality', () => {
-    it('should transform basic css with static object literal', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        
-        const styles = css({
-          color: 'red',
-          fontSize: 14,
-          backgroundColor: 'blue'
-        });
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _3 = "._qhg11u{background-color:blue}";
-        var _2 = "._1it30mr{font-size:14px}";
-        var _ = "._jpi694{color:red}";
-        var styles = null;
-        "
-      `);
-    });
-
-    it('should transform css prop with static object literal', async () => {
-      const code = `
-        import '@compiled/react';
-        
-        <div css={{
-          color: 'red',
-          fontSize: 14,
-          padding: '10px'
-        }}>Hello</div>
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React1 from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _3 = "._1ojmb8f{padding:10px}";
-        var _2 = "._1it30mr{font-size:14px}";
-        var _ = "._jpi694{color:red}";
-        /*#__PURE__*/ React.createElement("div", {
-            className: ax([
-                "_jpi694",
-                "_1it30mr",
-                "_1ojmb8f"
-            ])
-        }, "Hello");
-        "
-      `);
-    });
-
-    it('should transform css prop with string literal', async () => {
-      const code = `
-        import '@compiled/react';
-        
-        <div css="color: red; font-size: 14px;">Hello</div>
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React1 from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _2 = "._1elxrq9{font-size: 14px}";
-        var _ = "._1wszpi4{color: red}";
-        /*#__PURE__*/ React.createElement("div", {
-            className: ax([
-                "_1wszpi4",
-                "_1elxrq9"
-            ])
-        }, "Hello");
-        "
-      `);
-    });
-
-    it('should handle nested objects for pseudo-selectors', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        
-        const styles = css({
-          color: 'red',
-          ':hover': {
-            color: 'blue',
-            backgroundColor: 'yellow'
-          }
-        });
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _3 = "._1rldtxs:hover{background-color:yellow}";
-        var _2 = "._1yoobt8:hover{color:blue}";
-        var _ = "._jpi694{color:red}";
-        var styles = null;
-        "
-      `);
-    });
-
-    it('should handle media queries with static values', async () => {
-      const code = `
-        import '@compiled/react';
-        
-        <div css={{
-          color: 'red',
-          '@media screen': {
-            color: 'blue',
-            fontSize: 16
-          }
-        }}>Hello</div>
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React1 from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _3 = "@media screen{._y3kiq7{font-size:16px}}";
-        var _2 = "@media screen{._1yoobt8{color:blue}}";
-        var _ = "._jpi694{color:red}";
-        /*#__PURE__*/ React.createElement("div", {
-            className: ax([
-                "_jpi694",
-                "_1yoobt8",
-                "_y3kiq7"
-            ])
-        }, "Hello");
-        "
-      `);
-    });
+      });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _3 = "@media screen{._93hc8jm4{font-size:16px}}";
+      const _2 = "@media screen{._aooa2scz{color:blue}}";
+      const _ = "._3hisej0k{color:red}";
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisej0k",
+              "_aooa2scz",
+              "_93hc8jm4"
+          ])
+      });
+      "
+    `);
   });
 
-  describe('cssMap() - Supported Static Functionality', () => {
-    it('should transform basic cssMap with static values', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-
-        const styles = cssMap({
-          primary: {
-            color: 'red',
-            backgroundColor: 'blue'
-          },
-          secondary: {
-            color: 'green',
-            fontSize: 14
-          }
-        });
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _4 = "._1it30mr{font-size:14px}";
-        var _3 = "._1ulwugp{color:green}";
-        var _2 = "._qhg11u{background-color:blue}";
-        var _ = "._jpi694{color:red}";
-        var styles = {
-            primary: "_jpi694 _qhg11u",
-            secondary: "_1ulwugp _1it30mr"
-        };
-        "
-      `);
-    });
-
-    it('should transform cssMap usage in css prop', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-
-        const styles = cssMap({
-          danger: {
-            color: 'red',
-            backgroundColor: 'red'
-          },
-          success: {
-            color: 'green',
-            backgroundColor: 'green'
-          }
-        });
-
-        const Component = () => <div>
-          <span css={styles.danger} />
-          <span css={styles.success} />
-        </div>
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React1 from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _4 = "._uo4p87{background-color:green}";
-        var _3 = "._1ulwugp{color:green}";
-        var _2 = "._1u1ezxk{background-color:red}";
-        var _ = "._jpi694{color:red}";
-        var styles = {
-            danger: "_jpi694 _1u1ezxk",
-            success: "_1ulwugp _uo4p87"
-        };
-        var Component = function() {
-            return /*#__PURE__*/ React.createElement("div", null, /*#__PURE__*/ React.createElement("span", {
-                className: ax([
-                    styles.danger
-                ])
-            }), /*#__PURE__*/ React.createElement("span", {
-                className: ax([
-                    styles.success
-                ])
-            }));
-        };
-        "
-      `);
-    });
-
-    it('should handle empty cssMap variants', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-
-        const styles = cssMap({
-          empty: {},
-          filled: {
-            color: 'red'
-          }
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _ = "._jpi694{color:red}";
-        var styles = {
-            empty: "",
-            filled: "_jpi694"
-        };
-        "
-      `);
-    });
-
-    it('should handle cssMap with pseudo-selectors and media queries', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-
-        const styles = cssMap({
-          interactive: {
-            color: 'blue',
-            ':hover': {
-              color: 'red'
-            },
-            '@media screen': {
-              fontSize: 16
-            }
-          }
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _3 = "@media screen{._y3kiq7{font-size:16px}}";
-        var _2 = "._jpi694:hover{color:red}";
-        var _ = "._1yoobt8{color:blue}";
-        var styles = {
-            interactive: "_1yoobt8 _jpi694 _y3kiq7"
-        };
-        "
-      `);
-    });
-  });
-
-  describe('Strict Mode Restrictions - Should NOT Transform', () => {
-    it('should not transform css with dynamic variables', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        import color from './module';
-        
-        const styles = css({
-          color: color
-        });
-      `;
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should not transform css with function calls', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        
-        const styles = css({
-          color: getColor()
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should not transform css with binary expressions', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        
-        const styles = css({
-          width: 1 + 1 + "px"
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should not transform css with member expressions', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        
-        const theme = { colors: { primary: 'red' } };
-        const styles = css({
-          color: theme.colors.primary
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should not transform cssMap with dynamic values', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-        
-        const primaryColor = 'red';
-        const styles = cssMap({
-          primary: {
-            color: primaryColor
-          }
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should not transform cssMap with function calls', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-        
-        const styles = cssMap({
-          primary: {
-            color: getColor()
-          }
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should error on styled tagged template usage', async () => {
-      const code = `
-        import { styled } from '@compiled/react';
-        
-        const Button = styled.div\`
-          color: red;
-        \`;
-      `;
-      const error = await expectToThrow(code);
-      expectToContain(String(error), 'RuntimeError: unreachable');
-    });
-
-    it('should error on styled component factory calls', async () => {
-      const code = `
-        import { styled } from '@compiled/react';
-        
-        const Button = styled('div')({
-          color: 'red'
-        });
-      `;
-      const error = await expectToThrow(code);
-      expectToContain(String(error), 'RuntimeError: unreachable');
-    });
-  });
-
-  describe('Edge Cases', () => {
-    it('should handle mixed static and dynamic (only transform static)', async () => {
-      // Test that static parts work while dynamic parts are ignored
-      const code = `
-        import '@compiled/react';
-        
-        const dynamicColor = 'red';
-        <div css={{
-          color: 'blue',        // static - should work
-          fontSize: 14,         // static - should work  
-          backgroundColor: dynamicColor  // dynamic - should not work
-        }}>Hello</div>
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-
-    it('should handle cssMap with mixed static and dynamic variants', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-        
-        const dynamicColor = 'red';
-        const styles = cssMap({
-          static: {
-            color: 'blue',
-            fontSize: 14
-          },
-          dynamic: {
-            color: dynamicColor
-          }
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "Error: failed to handle: failed to invoke plugin: failed to invoke plugin on 'Some("test.tsx")'
-
-        Caused by:
-            0: failed to invoke \`/Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm\` as js transform plugin at /Users/sjackson3/Documents/atlassian/compiled-perf-fork/packages/swc-plugin2/compiled_swc_plugin2.wasm
-            1: RuntimeError: unreachable"
-      `);
-    });
-  });
-
-  describe('Comparison with Babel Plugin Behavior', () => {
-    it('should produce same output as babel plugin for basic css', async () => {
-      const code = `
-        import { css } from '@compiled/react';
-        
-        const styles = css({
-          color: 'red',
-          fontSize: 14
-        });
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _2 = "._1it30mr{font-size:14px}";
-        var _ = "._jpi694{color:red}";
-        var styles = null;
-        "
-      `);
-    });
-
-    it('should produce same output as babel plugin for basic cssMap', async () => {
-      const code = `
-        import { cssMap } from '@compiled/react';
-
-        const styles = cssMap({
-          danger: {
-            color: 'red',
-            backgroundColor: 'red'
-          },
-          success: {
-            color: 'green',
-            backgroundColor: 'green'
-          }
-        });
-
-        const Component = () => <div>
-          <span css={styles.danger} />
-          <span css={styles.success} />
-        </div>
-      `;
-
-      const out = await transformResultString(code);
-      expect(out).toMatchInlineSnapshot(`
-        "import * as React1 from "react";
-        import { ax, ix, CC, CS } from "@compiled/react/runtime";
-        var _4 = "._uo4p87{background-color:green}";
-        var _3 = "._1ulwugp{color:green}";
-        var _2 = "._1u1ezxk{background-color:red}";
-        var _ = "._jpi694{color:red}";
-        var styles = {
-            danger: "_jpi694 _1u1ezxk",
-            success: "_1ulwugp _uo4p87"
-        };
-        var Component = function() {
-            return /*#__PURE__*/ React.createElement("div", null, /*#__PURE__*/ React.createElement("span", {
-                className: ax([
-                    styles.danger
-                ])
-            }), /*#__PURE__*/ React.createElement("span", {
-                className: ax([
-                    styles.success
-                ])
-            }));
-        };
-        "
-      `);
-    });
-
-    it('should match Babel sorting: base, pseudo (LVFHA), then at-rules', async () => {
-      const code = `
-        import '@compiled/react';
-        <div css={{
-          color: 'black',
-          ':focus': { color: 'green' },
-          ':hover': { color: 'blue' },
-          ':active': { color: 'purple' },
+  it('cssMap with @media in variants', async () => {
+    const code = `
+      import { cssMap } from '@compiled/react';
+      const variants = cssMap({
+        primary: {
           '@media screen': { color: 'red' },
-        }} />
-      `;
-      const out = await transformResultString(code);
-      // In emitted code, declarations are inserted in reverse, so we expect:
-      // @media block first, then :focus, then :hover, then :active, then base rule last.
-      const baseIdx = out.indexOf('{color:black}');
-      const focusIdx = out.indexOf(':focus{color:green}');
-      const hoverIdx = out.indexOf(':hover{color:blue}');
-      const activeIdx = out.indexOf(':active{color:purple}');
-      const mediaIdx = out.indexOf('@media screen{');
+        },
+        secondary: {
+          '@media screen': { fontSize: 16 },
+        },
+      });
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _2 = "@media screen{._93hc8jm4{font-size:16px}}";
+      const _ = "@media screen{._aooaej0k{color:red}}";
+      const variants = {
+          primary: "_aooaej0k",
+          secondary: "_93hc8jm4"
+      };
+      "
+    `);
+  });
 
-      expect(mediaIdx).toBeGreaterThan(-1);
-      expect(focusIdx).toBeGreaterThan(mediaIdx);
-      expect(hoverIdx).toBeGreaterThan(focusIdx);
-      expect(activeIdx).toBeGreaterThan(hoverIdx);
-      expect(baseIdx).toBeGreaterThan(activeIdx);
-    });
+  it('numeric values add px where required and preserve unitless', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        fontSize: 16,
+        lineHeight: 2,
+        marginTop: 0,
+      });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _3 = "._5mpx4k5u{margin-top:0}";
+      const _2 = "._6hxr7ngg{line-height:2}";
+      const _ = "._7ql18jm4{font-size:16px}";
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_7ql18jm4",
+              "_6hxr7ngg",
+              "_5mpx4k5u"
+          ])
+      });
+      "
+    `);
+  });
 
-    it('should emit atomic CSS and deduplicate identical declarations', async () => {
-      const code = `
-        import '@compiled/react';
-        <> 
-          <div css={{ color: 'black', fontSize: 12 }} />
-          <div css={{ color: 'black', fontSize: 12 }} />
-        </>
-      `;
+  it('padding shorthand expands to longhands even without explicit longhands', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        padding: '10px 20px',
+      });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _4 = "._7wlm154n{padding-left:20px}";
+      const _3 = "._da1ibjn1{padding-bottom:10px}";
+      const _2 = "._vhdx154n{padding-right:20px}";
+      const _ = "._8edybjn1{padding-top:10px}";
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_8edybjn1",
+              "_vhdx154n",
+              "_da1ibjn1",
+              "_7wlm154n"
+          ])
+      });
+      "
+    `);
+  });
 
-      // Transform with SWC
-      const swcOut = await transformResultString(code);
+  it('padding shorthand expanded when longhands present', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const styles = css({
+        padding: '10px 20px',
+        paddingLeft: '4px',
+      });
+      <div css={styles} />
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _5 = "._7wlm9z5l{padding-left:4px}";
+      const _4 = "._7wlm154n{padding-left:20px}";
+      const _3 = "._da1ibjn1{padding-bottom:10px}";
+      const _2 = "._vhdx154n{padding-right:20px}";
+      const _ = "._8edybjn1{padding-top:10px}";
+      const styles = null;
+      /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_8edybjn1",
+              "_vhdx154n",
+              "_da1ibjn1",
+              "_7wlm154n",
+              "_7wlm9z5l"
+          ])
+      });
+      "
+    `);
+  });
 
-      // Basic atomic shape expectations (no combined rule)
-      expectNotToContain(swcOut, '{color:black;font-size:12px}');
-      expectToContain(swcOut, ['{color:black}', '{font-size:12px}']);
+  it('styled basic object call (extract:true) without CC/CS', async () => {
+    const code = `
+      import { styled } from '@compiled/react';
+      const Button = styled.div({ color: 'red' });
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { forwardRef } from "react";
+      import { ax, ix } from "@compiled/react/runtime";
+      const _ = "._3hisej0k{color:red}";
+      import { styled } from '@compiled/react';
+      const Button = /*#__PURE__*/ forwardRef(({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr)=>{
+          if (__cmplp.innerRef) {
+              throw new Error("Please use 'ref' instead of 'innerRef'.");
+          }
+          return /*#__PURE__*/ React.createElement(C, {
+              className: ax([
+                  "_3hisej0k",
+                  __cmplp.className
+              ]),
+              style: {
+                  ...__cmpls
+              },
+              ref: __cmplr,
+              ...__cmplp
+          });
+      });
+      "
+    `);
+  });
 
-      // Count occurrences for SWC
-      const swcColorCount = (swcOut.match(/\{color:black\}/g) || []).length;
-      const swcFontSizeCount = (swcOut.match(/\{font-size:12px\}/g) || []).length;
+  it('styled basic object call (extract:false) wraps with CC/CS', async () => {
+    const code = `
+      import { styled } from '@compiled/react';
+      const Button = styled.div({ color: 'red' });
+    `;
+    const out = await transformResultString(code, { extract: false });
+    expect(out).toMatchInlineSnapshot(`
+      "import { forwardRef } from "react";
+      import { ax, ix, CC, CS } from "@compiled/react/runtime";
+      const _ = "._3hisej0k{color:red}";
+      import { styled } from '@compiled/react';
+      const Button = /*#__PURE__*/ forwardRef(({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr)=>{
+          if (__cmplp.innerRef) {
+              throw new Error("Please use 'ref' instead of 'innerRef'.");
+          }
+          return /*#__PURE__*/ React.createElement(CC, null, /*#__PURE__*/ React.createElement(CS, null, [
+              _
+          ]), /*#__PURE__*/ React.createElement(C, {
+              className: ax([
+                  "_3hisej0k",
+                  __cmplp.className
+              ]),
+              style: {
+                  ...__cmpls
+              },
+              ref: __cmplr,
+              ...__cmplp
+          }));
+      });
+      "
+    `);
+  });
 
-      // SWC should dedupe: each unique declaration should appear only once
-      expect(swcColorCount).toBe(1);
-      expect(swcFontSizeCount).toBe(1);
-    });
-
-    it('should sort border shorthand before border-top before border-top-color', async () => {
-      const code = `
-        import '@compiled/react';
-        <div css={{
-          borderTopColor: 'red',
-          borderTop: '1px solid',
-          border: 'none',
-        }} />
-      `;
-      const out = await transformResultString(code);
-      const idxBorder = out.indexOf('{border:none}');
-      const idxBorderTop = out.indexOf('{border-top:1px solid}');
-      const idxBorderTopColor = out.indexOf('{border-top-color:red}');
-      expect(idxBorder).toBeGreaterThan(-1);
-      expect(idxBorderTop).toBeGreaterThan(-1);
-      expect(idxBorderTopColor).toBeGreaterThan(-1);
-      // For SWC emission order, the last inserted appears last in output text.
-      // We only assert all exist; order may differ due to insertion mechanics.
-      expect(idxBorder).toBeGreaterThan(-1);
-    });
-
-    it('should expand padding shorthand into longhands and include explicit longhands', async () => {
-      const code = `
-        import '@compiled/react';
-        <div css={{
-          paddingLeft: 10,
-          padding: '10px 20px',
-        }} />
-      `;
-      const out = await transformResultString(code);
-      // Expect expanded longhands from padding shorthand
-      expectToContain(out, [
-        '{padding-top:10px}',
-        '{padding-right:20px}',
-        '{padding-bottom:10px}',
-        '{padding-left:20px}',
-      ]);
-      // And explicit longhand should also be present
-      expectToContain(out, '{padding-left:10px}');
-    });
-
-    it('should include !important in value hashing for pseudos', async () => {
-      const code = `
-        import '@compiled/react';
-        <div css={{
-          ':hover': { color: 'blue !important' }
-        }} />
-      `;
-
-      const out = await transformResultString(code);
-      // Expect :hover rule with color:blue !important and a stable class name format
-      expectToContain(out, [':hover{color:blue !important}']);
-    });
+  it('deduplicates identical styles across multiple elements', async () => {
+    const code = `
+      import { css } from '@compiled/react';
+      const a = css({ color: 'red' });
+      const b = css({ color: 'red' });
+      <>
+        <div css={[a]} />
+        <div css={[b]} />
+      </>
+    `;
+    const out = await transformResultString(code, { extract: true });
+    expect(out).toMatchInlineSnapshot(`
+      "import { ax, ix } from "@compiled/react/runtime";
+      const _ = "._3hisej0k{color:red}";
+      const a = null;
+      const b = null;
+      /*#__PURE__*/ React.createElement(React.Fragment, null, /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisej0k"
+          ])
+      }), /*#__PURE__*/ React.createElement("div", {
+          className: ax([
+              "_3hisej0k"
+          ])
+      }));
+      "
+    `);
   });
 });
