+++ packages/swc-plugin/tests/swc-output.js
@@ -1,70 +1,73 @@
+const fs = require('fs');
 const path = require('path');
 
 const swc = require('@swc/core');
 
-/**
- * Transform code using our SWC plugin
- */
-async function transform(code, options = {}) {
-  const wasmPath = path.join(__dirname, '..', 'compiled_swc_plugin.wasm');
+const lib = require('..');
 
+async function transform(code, options = {}) {
+  const wasmPath = path.join(__dirname, '..', 'compiled_swc_plugin2.wasm');
   const pluginOptions = {
-    processXcss: true,
     importSources: ['@compiled/react'],
-    strictMode: true, // Enable strict mode
+    extract: true,
     ...options,
   };
+  // Only enable the plugin when the code includes configured import sources
+  const hasCssLikeJsx = /\b(?:css|[A-Za-z]*xcss)\s*=/i.test(code);
+  const hasApiCalls = /(cssMap|styled|keyframes)\s*\(/.test(code);
+  const hasImports = lib.shouldEnableForCode2
+    ? lib.shouldEnableForCode2(code, pluginOptions)
+    : pluginOptions.importSources.some((s) => code.includes(s));
+  const enablePlugin =
+    options.forceEnable !== undefined
+      ? !!options.forceEnable
+      : hasImports || hasCssLikeJsx || hasApiCalls;
 
-  try {
-    const result = await swc.transform(code, {
-      filename: 'test.tsx',
-      jsc: {
-        parser: {
-          syntax: 'typescript',
-          tsx: true,
-        },
-        experimental: {
-          plugins: [[wasmPath, pluginOptions]],
-        },
-      },
-    });
-
-    return result.code;
-  } catch (error) {
-    throw error;
-  }
+  const swcOptions = {
+    filename: 'test.tsx',
+    cwd: process.cwd(),
+    jsc: {
+      target: 'es2020',
+      externalHelpers: true,
+      parser: { syntax: 'typescript', tsx: true },
+      experimental: enablePlugin
+        ? {
+            plugins: [
+              [
+                wasmPath,
+                (() => {
+                  const extra =
+                    pluginOptions && pluginOptions.extractStylesToDirectory
+                      ? {
+                          filename: 'test.tsx',
+                          sourceFileName: '../src/test.tsx',
+                        }
+                      : {};
+                  return { ...pluginOptions, ...extra };
+                })(),
+              ],
+            ],
+          }
+        : {},
+    },
+  };
+  const result = await swc.transform(code, swcOptions);
+  return result.code;
 }
 
-// Helper to always return a string snapshot: transformed code or error string
 async function transformResultString(code, options = {}) {
   try {
-    const output = await transform(code, options);
-    return normalizeOutput(output);
-  } catch (error) {
-    return normalizeOutput(String(error));
+    return normalizeOutput(await transform(code, options));
+  } catch (e) {
+    return normalizeOutput(String(e));
   }
 }
 
 function normalizeOutput(str) {
   return str.replace(
-    /packages\/swc-plugin\/compiled_swc_plugin\.wasm/g,
-    'packages/swc-plugin/compiled_swc_plugin.wasm'
+    /packages\/swc-plugin2\/compiled_swc_plugin2\.wasm/g,
+    'packages/swc-plugin/compiled_swc_plugin2.wasm'
   );
 }
 
-const code = `
-  import { css } from '@compiled/react';
-
-  const styles = css({
-    color: 'red',
-    '& + label ~ div': {
-        color: 'blue'
-    },
-  });
-
-  <div
-    css={[styles]}
-  />
-`;
-const out = await transformResultString(code);
-console.log(out);
+module.exports = { transform, transformResultString };
