--- packages/swc-plugin2/.gitignore
@@ -1,5 +1,0 @@
-dist
-node_modules
-compiled_swc_plugin2.wasm
-target
-
