--- packages/swc-plugin2/package.json
@@ -1,62 +1,0 @@
-{
-  "name": "@compiled/swc-plugin2",
-  "version": "0.0.1",
-  "description": "Experimental SWC plugin for @compiled/react focusing on css and cssMap with extraction.",
-  "keywords": [
-    "swc",
-    "plugin",
-    "css-in-js",
-    "compiled",
-    "react",
-    "wasm",
-    "rust"
-  ],
-  "homepage": "https://compiledcssinjs.com/docs/pkg-swc-plugin2",
-  "bugs": "https://github.com/atlassian-labs/compiled/issues/new?assignees=&labels=bug&template=bug_report.md",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/atlassian-labs/compiled.git",
-    "directory": "packages/swc-plugin2"
-  },
-  "license": "Apache-2.0",
-  "author": "Shanon Jackson",
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist",
-    "compiled_swc_plugin2.wasm",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "npm run build:rust && npm run build:ts && npm run copy-wasm",
-    "build:dev": "cargo build --target wasm32-wasip1",
-    "build:rust": "cargo build --target wasm32-wasip1 --release",
-    "build:ts": "tsc",
-    "check": "cargo check",
-    "clean": "cargo clean && rm -rf dist compiled_swc_plugin2.wasm",
-    "clippy": "cargo clippy",
-    "copy-wasm": "cp target/wasm32-wasip1/release/compiled_swc_plugin2.wasm ./compiled_swc_plugin2.wasm",
-    "fmt": "cargo fmt",
-    "prepare": "npm run build",
-    "pretest": "npm run build:rust && npm run copy-wasm",
-    "test": "jest --no-cache --config ./jest.config.js",
-    "test:ts": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "@swc/core": "1.3.107",
-    "@types/node": "^20.0.0",
-    "jest-extended": "^4.0.2",
-    "typescript": "^5.0.0"
-  },
-  "peerDependencies": {
-    "@swc/core": ">=1.3.0",
-    "react": ">=16.8.0"
-  },
-  "engines": {
-    "node": ">=16.0.0"
-  },
-  "swc": {
-    "plugin": "./target/wasm32-wasip1/release/compiled_swc_plugin2.wasm"
-  }
-}
