--- packages/swc-plugin2/src/utils/mod.rs
@@ -1,4 +1,0 @@
-pub mod ast;
-pub mod css_builder;
-pub mod xcss;
-
