--- packages/swc-plugin2/src/visitors/css.rs
@@ -1,173 +1,0 @@
-use crate::types::TransformState;
-use crate::utils::{css_builder, ast};
-use crate::types::KeyframeVarSpec;
-use swc_core::{
-    common::DUMMY_SP,
-    ecma::ast::*,
-};
-
-pub fn is_css_call(call: &CallExpr, state: &TransformState) -> bool {
-    if let Some(ref imports) = state.compiled_imports {
-        if let Some(ref css_imports) = imports.css {
-            match &call.callee {
-                Callee::Expr(expr) => match expr.as_ref() { Expr::Ident(ident) => css_imports.contains(&ident.sym.to_string()), _ => false },
-                _ => false,
-            }
-        } else { false }
-    } else {
-        match &call.callee { Callee::Expr(expr) => match expr.as_ref() { Expr::Ident(ident) => ident.sym.as_ref() == "css", _ => false }, _ => false }
-    }
-}
-
-pub fn visit_css_prop_jsx_element(
-    opening_elem: &mut JSXOpeningElement,
-    state: &mut TransformState,
-    css_content_to_var: &mut std::collections::HashMap<String, String>,
-    collected_css_sheets: &mut Vec<(String, String)>,
-    extract: bool,
-) -> bool {
-    let mut css_attr_index = None;
-    for (i, attr) in opening_elem.attrs.iter().enumerate() {
-        if let JSXAttrOrSpread::JSXAttr(jsx_attr) = attr {
-            if let JSXAttrName::Ident(ident) = &jsx_attr.name { if ident.sym.as_ref() == "css" { css_attr_index = Some(i); break; } }
-        }
-    }
-    let css_attr_index = match css_attr_index { Some(index) => index, None => return false };
-
-    let is_string_literal = match &opening_elem.attrs[css_attr_index] { JSXAttrOrSpread::JSXAttr(attr) => matches!(&attr.value, Some(JSXAttrValue::Lit(Lit::Str(_)))), _ => false };
-    if is_string_literal {
-        if let JSXAttrOrSpread::JSXAttr(attr) = &opening_elem.attrs[css_attr_index] {
-            let str_lit = match &attr.value { Some(JSXAttrValue::Lit(Lit::Str(s))) => s.clone(), _ => unreachable!() };
-            let rules = css_builder::build_atomic_rules_from_expression_with_state(&Expr::Lit(Lit::Str(str_lit.clone())), state);
-            if !rules.is_empty() {
-                let (sheets, class_names) = css_builder::transform_atomic_rules_to_sheets(&rules);
-                for sheet in sheets { let _ = add_css_sheet_with_deduplication(css_content_to_var, collected_css_sheets, &sheet); }
-                if extract && !class_names.is_empty() {
-                    if let JSXAttrOrSpread::JSXAttr(attr_mut) = &mut opening_elem.attrs[css_attr_index] {
-                        attr_mut.name = JSXAttrName::Ident(ast::create_ident("className"));
-                        let ax_call = create_ax_call(&class_names);
-                        attr_mut.value = Some(JSXAttrValue::JSXExprContainer(JSXExprContainer { span: DUMMY_SP, expr: JSXExpr::Expr(Box::new(ax_call)) }));
-                    }
-                } else if !extract && !class_names.is_empty() {
-                    // Wrap inside <CC><CS>{[...]}<div .../></CC>
-                    return wrap_parent_jsx_with_cc_cs(opening_elem, &class_names, &[]);
-                } else { opening_elem.attrs.remove(css_attr_index); }
-                return true;
-            }
-        }
-    }
-
-    // Expression
-    let binding: Option<JSXAttrValue> = match &opening_elem.attrs[css_attr_index] { JSXAttrOrSpread::JSXAttr(attr) => attr.value.clone(), _ => None };
-    let expr = match get_jsx_attribute_expression(&binding) { Some(expr) => expr, None => return false };
-
-    // Handle arrays like css={[styles, other]}
-    if let Expr::Array(array_lit) = expr {
-        let mut all_classes: Vec<String> = Vec::new();
-        let mut all_sheet_vars: Vec<String> = Vec::new();
-        for maybe_elem in &array_lit.elems {
-            if let Some(ExprOrSpread { expr: item_expr, .. }) = maybe_elem {
-                if let Expr::Ident(id) = item_expr.as_ref() {
-                    if let Some(info) = state.css_classes_by_ident.get(&id.sym.to_string()) {
-                        all_classes.extend(info.class_names.clone());
-                        all_sheet_vars.extend(info.sheet_var_names.clone());
-                    }
-                }
-            }
-        }
-        if extract && !all_classes.is_empty() {
-            let combined = all_classes.join(" ");
-            let ax_call = Expr::Call(CallExpr { span: DUMMY_SP, callee: Callee::Expr(Box::new(Expr::Ident(ast::create_ident("ax")))), args: vec![ExprOrSpread { spread: None, expr: Box::new(Expr::Array(ArrayLit { span: DUMMY_SP, elems: vec![Some(ExprOrSpread { spread: None, expr: Box::new(Expr::Lit(Lit::Str(ast::create_str_lit(&combined)))) })] })) }], type_args: None });
-            if let JSXAttrOrSpread::JSXAttr(attr_mut) = &mut opening_elem.attrs[css_attr_index] {
-                attr_mut.name = JSXAttrName::Ident(ast::create_ident("className"));
-                attr_mut.value = Some(JSXAttrValue::JSXExprContainer(JSXExprContainer { span: DUMMY_SP, expr: JSXExpr::Expr(Box::new(ax_call)) }));
-            }
-            return true;
-        } else if !extract && !all_classes.is_empty() {
-            return wrap_parent_jsx_with_cc_cs(opening_elem, &all_classes, &all_sheet_vars);
-        }
-    }
-
-    // If css prop references an identifier produced by css(), rewrite to className with ax([...classes])
-    if let Expr::Ident(id) = expr {
-        if let Some(classes) = state.css_classes_by_ident.get(&id.sym.to_string()) {
-            if extract {
-                let elems: Vec<Option<ExprOrSpread>> = classes
-                    .class_names
-                    .iter()
-                    .map(|cn| Some(ExprOrSpread { spread: None, expr: Box::new(Expr::Lit(Lit::Str(ast::create_str_lit(cn)))) }))
-                    .collect();
-                let ax_call = Expr::Call(CallExpr { span: DUMMY_SP, callee: Callee::Expr(Box::new(Expr::Ident(ast::create_ident("ax")))), args: vec![ExprOrSpread { spread: None, expr: Box::new(Expr::Array(ArrayLit { span: DUMMY_SP, elems })) }], type_args: None });
-                if let JSXAttrOrSpread::JSXAttr(attr_mut) = &mut opening_elem.attrs[css_attr_index] {
-                    attr_mut.name = JSXAttrName::Ident(ast::create_ident("className"));
-                    attr_mut.value = Some(JSXAttrValue::JSXExprContainer(JSXExprContainer { span: DUMMY_SP, expr: JSXExpr::Expr(Box::new(ax_call)) }));
-                }
-                return true;
-            } else {
-                return wrap_parent_jsx_with_cc_cs(opening_elem, &classes.class_names, &classes.sheet_var_names);
-            }
-        }
-    }
-
-    let should_rewrite_to_classname = matches!(expr, Expr::Member(_) | Expr::Ident(_));
-    if should_rewrite_to_classname {
-        let ax_call = Expr::Call(CallExpr { span: DUMMY_SP, callee: Callee::Expr(Box::new(Expr::Ident(ast::create_ident("ax")))), args: vec![ExprOrSpread { spread: None, expr: Box::new(Expr::Array(ArrayLit { span: DUMMY_SP, elems: vec![Some(ExprOrSpread { spread: None, expr: Box::new(expr.clone()) })] })) }], type_args: None });
-        if let JSXAttrOrSpread::JSXAttr(attr_mut) = &mut opening_elem.attrs[css_attr_index] {
-            attr_mut.name = JSXAttrName::Ident(ast::create_ident("className"));
-            attr_mut.value = Some(JSXAttrValue::JSXExprContainer(JSXExprContainer { span: DUMMY_SP, expr: JSXExpr::Expr(Box::new(ax_call)) }));
-        }
-        return true;
-    }
-
-    let atomic_rules = css_builder::build_atomic_rules_from_expression_with_state(expr, state);
-    if !atomic_rules.is_empty() {
-        let (sheets, class_names) = css_builder::transform_atomic_rules_to_sheets(&atomic_rules);
-        for sheet in sheets { let _ = add_css_sheet_with_deduplication(css_content_to_var, collected_css_sheets, &sheet); }
-        if extract && !class_names.is_empty() {
-            if let JSXAttrOrSpread::JSXAttr(attr) = &mut opening_elem.attrs[css_attr_index] {
-                attr.name = JSXAttrName::Ident(ast::create_ident("className"));
-                let ax_call = create_ax_call(&class_names);
-                attr.value = Some(JSXAttrValue::JSXExprContainer(JSXExprContainer { span: DUMMY_SP, expr: JSXExpr::Expr(Box::new(ax_call)) }));
-            }
-        } else if !extract && !class_names.is_empty() {
-            // No sheet var names tracked here (inline object); just use class names
-            return wrap_parent_jsx_with_cc_cs(opening_elem, &class_names, &[]);
-        } else { opening_elem.attrs.remove(css_attr_index); }
-        return true;
-    }
-    false
-}
-
-fn get_jsx_attribute_expression(value: &Option<JSXAttrValue>) -> Option<&Expr> {
-    match value {
-        Some(JSXAttrValue::JSXExprContainer(container)) => match &container.expr { JSXExpr::Expr(expr) => Some(expr.as_ref()), _ => None },
-        _ => None,
-    }
-}
-
-fn create_ax_call(class_names: &[String]) -> Expr {
-    let array_elements: Vec<Option<ExprOrSpread>> = class_names
-        .iter()
-        .map(|class_name| Some(ExprOrSpread { spread: None, expr: Box::new(Expr::Lit(Lit::Str(ast::create_str_lit(class_name)))) }))
-        .collect();
-    Expr::Call(CallExpr { span: DUMMY_SP, callee: Callee::Expr(Box::new(Expr::Ident(ast::create_ident("ax")))), args: vec![ExprOrSpread { spread: None, expr: Box::new(Expr::Array(ArrayLit { span: DUMMY_SP, elems: array_elements })) }], type_args: None })
-}
-
-fn add_css_sheet_with_deduplication(
-    css_content_to_var: &mut std::collections::HashMap<String, String>,
-    collected_css_sheets: &mut Vec<(String, String)>,
-    css_content: &str,
-) -> String {
-    if let Some(existing_var_name) = css_content_to_var.get(css_content) { return existing_var_name.clone(); }
-    let index = collected_css_sheets.len();
-    let var_name = if index == 0 { "_".to_string() } else { format!("_{}", index + 1) };
-    css_content_to_var.insert(css_content.to_string(), var_name.clone());
-    collected_css_sheets.push((var_name.clone(), css_content.to_string()));
-    var_name
-}
-
-fn wrap_parent_jsx_with_cc_cs(_opening_elem: &mut JSXOpeningElement, _class_names: &[String], _sheet_var_names: &[String]) -> bool {
-    // TODO: Implement full CC/CS wrapper emission (JSX tree rewrite). For now, noop to keep tests green.
-    true
-}
-
